requirements.txt
faiss-cpu
sentence-transformers==2.6.1
transformers==4.41.2
huggingface_hub==0.23.4
openai
tiktoken
requests